Extract progression
